backend/requirements.txt
django
djangorestframework
random-password-generator
django-cors-headers
pip install drf-yasgdrf-yasg